etl/requirements.txt
backoff==1.11.1
clickhouse-driver==0.2.3
confluent-kafka==1.7.0
pydantic==1.9.0
python-dotenv==0.19.0
pyyaml==6.0.0
redis==4.1.4
